% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_goodnessOfFit.R
\name{calc_goodnessOfFit}
\alias{calc_goodnessOfFit}
\title{calc_GoodnessOfFit}
\usage{
calc_goodnessOfFit(predicted, observed)
}
\arguments{
\item{predicted}{vector of predicted values}

\item{observed}{vector of observed values}
}
\value{
A list with the following items:\cr
       \describe{
          \item{\code{rmse}}{Root mean square error 
                      = \code{sqrt(mean((predicted - observed)^2), na.rm = TRUE))}}
          \item{\code{rSquared}}{Coefficient of determination =  
                           \code{1 - sum((predicted - observed)^2) / sum((observed - mean(observed, na.rm = TRUE))^2)} 
                           Useful indicator, although for non-linear models it is not strictly appropriate since the 
                           'null hypothesis', a horizontal line, is not necessarily a subset of the model space. 
                           Based on the traditional definition, see \code{?caret::R2}.}
          \item{\code{normalityTestResiduals}}{P-value of a kolmogorov-smirnoff test 
                                         (\code{\link[stats]{ks.test}}). Null-hypothesis: Normal distribution 
                                          with mean 0 and standard deviation = sd(predicted - observed))}
        }
}
\description{
Function to calculate goodness of fit measures for regressions.
}
\examples{
test <- calc_goodnessOfFit(predicted = 1:10 + rnorm(10,0,0.3), observed = 1:10)
}
\seealso{
\code{\link{plot_goodnessOfFit}}
}
