% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret2lstmInput.R
\name{caret2lstmInput}
\alias{caret2lstmInput}
\title{caret2lstminput}
\usage{
caret2lstmInput(dat)
}
\arguments{
\item{dat}{data.frame as returned by \code{\link{lstmInput2caret}}}
}
\value{
a list of arrays as required by \code{\link{fitLSTMmodel}} and \code{\link{mxLSTM}}
        Same structure as the return value of \code{\link{transformLSTMinput}}
}
\description{
transforms a data.frame as returned by \code{\link{lstmInput2caret}} into a list of arrays as required by \code{\link{mxLSTM}}
}
