% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret2lstmInput1.R
\name{caret2lstmInput1}
\alias{caret2lstmInput1}
\title{caret2lstminput1}
\usage{
caret2lstmInput1(dat)
}
\arguments{
\item{dat}{data.frame as returned by \code{\link{lstmInput2caret}}}
}
\value{
data.table with one column per variable, one column "eventId", and one column "sequenceId"
}
\description{
transforms a data.frame as returned by \code{\link{lstmInput2caret}} 
             into a data.frame with all measurements of the same variable in one column.
             For preProcessing.
}
