% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLSTMmodel.R
\name{getLSTMmodel}
\alias{getLSTMmodel}
\title{getLSTMmodel}
\usage{
getLSTMmodel()
}
\value{
A list of functions similar to the output of caret's \code{\link[caret]{getModelInfo}}:\cr
}
\description{
Constructs a custom \code{\link{mxLSTM}} model for use in the caret 
\code{\link[caret]{train}} logic. It behaves slightly different than the usual 
caret models as retrieved by \code{\link[caret]{getModelInfo}}. See details.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{saveCaretLstmModel}}, \code{\link{loadCaretLstmModel}},
\code{\link{plot_trainHistory}}, \code{\link{fitLSTMmodel}}, 
\code{\link{predictLSTMmodel}}, \code{\link{getPreProcessor}}, 
\code{\link{predictPreProcessor}}, \code{\link{invertPreProcessor}}
}
