% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPreProcessor.R
\name{getPreProcessor}
\alias{getPreProcessor}
\title{getPreProcessor}
\usage{
getPreProcessor(dat, preProcess, preProcOptions = list(thresh = 0.95, ICAcomp
  = 3, k = 5, freqCut = 95/5, uniqueCut = 10, cutoff = 0.9))
}
\arguments{
\item{dat}{data.frame containing the data to be preprocessed.}

\item{preProcess}{Character vector with preprocessing steps for dat. 
Can be anything that caret allows 
(see \code{\link[caret]{preProcess}}) 
plus \code{c('scaleAgain', 'centerAgain', 'minMax')}.
\code{scaleAgain} and \code{centerAgain} are used to rescale 
data after a PCA or a similar transformation has 
destroyed the original scaling by caret.
minMax transforms the output into [0;1]. 
minMax is always executed as the last step.}

\item{preProcOptions}{list of options for the different preProcessing steps from caret.}
}
\value{
List with the following items:
        \itemize{
          \item{preProcessStep1:}{ preProcessing result for dat, 
          first step (all native caret methods). See \code{\link[caret]{preProcess}}}
          \item{preProcessStep2:}{ preProcessing result for dat for \code{scaleAgain, centerAgain}}
                               Result of caret preProcess with scale and center
          \item{preProcessMinMax:}{ preProcessing result for minMax trafo. 
                                  Comprises min value, max value and trafo function.}
        }
}
\description{
Preprocessing for regression in caret
}
\seealso{
\code{\link{predictPreProcessor}}, \code{\link{invertPreProcessor}}
}
