% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainHistory.mxLSTM.R
\name{getTrainHistory.mxLSTM}
\alias{getTrainHistory.mxLSTM}
\title{getTrainHistory.mxLSTM}
\usage{
\method{getTrainHistory}{mxLSTM}(model)
}
\arguments{
\item{model}{the model as returned by \code{\link{fitLSTMmodel}} or \code{\link{mxLSTM}}}
}
\value{
data.frame with train and eval performance
}
\description{
returns the logged performance metrics of an mxLSTM object.
}
