% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrainHistory.R
\name{getTrainHistory}
\alias{getTrainHistory}
\title{getTrainHistory}
\usage{
getTrainHistory(model)
}
\arguments{
\item{model}{the model}
}
\value{
data.frame with two columns: train and test with 
        performances on training and eval sets
}
\description{
gets the logged training performance of models
}
