% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertPreProcessor.R
\name{invertPreProcessor}
\alias{invertPreProcessor}
\title{invertPreProcessor}
\usage{
invertPreProcessor(preProcessor, dat)
}
\arguments{
\item{preProcessor}{Stored preProcessing procedure. Result of \code{\link{getPreProcessor}}.}

\item{dat}{data.frame containing the data to be preprocessed. Must have the correct column names.}
}
\value{
data.frame with data transofrmed by inverse preProcessing procedure
}
\description{
Do inverted transformation of stored preProcessing procedure on new data
}
\seealso{
\code{\link{getPreProcessor}}, \code{\link{predictPreProcessor}}
}
