% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCaretLstmModel.R
\name{loadCaretLstmModel}
\alias{loadCaretLstmModel}
\title{loadCaretLstmModel}
\usage{
loadCaretLstmModel(folder)
}
\arguments{
\item{folder}{The folder where the model was stored.}
}
\value{
caret training output with finalModel of class mxLSTM model
}
\description{
loads an LSTM model that was trained by caret and saved using \code{\link{saveCaretLstmModel}}.
}
\seealso{
\code{\link{loadLstmModel}}, \code{\link{saveCaretLstmModel}}
}
