% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLstmModel.R
\name{loadLstmModel}
\alias{loadLstmModel}
\title{loadLstmModel}
\usage{
loadLstmModel(folder)
}
\arguments{
\item{folder}{The folder where the model was stored.}
}
\value{
mxLSTM model
}
\description{
loads an LSTM mdoel that was saved using \code{\link{saveLstmModel}}.
}
\seealso{
\code{\link{saveLstmModel}}
}
