% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxLSTMsource.R
\name{lstmCell}
\alias{lstmCell}
\title{lstmCell}
\usage{
lstmCell(num.hidden, indata, prev.state, param, seqidx, layeridx, dropout = 0,
  zoneout = 0, batchNorm = FALSE, activation = "relu")
}
\arguments{
\item{num.hidden}{number of hidden neurons in the cell's state}

\item{indata}{the input to the cell}

\item{prev.state}{the memorized state from the previous cell.}

\item{param}{list of variables with weights and biases. must contain elements i2h.weight, i2h.bias, h2h.weight, h2h.bias}

\item{seqidx}{sequence index. Purely bookkeeping}

\item{layeridx}{index for the layer that the cell belongs to}

\item{dropout}{see \code{\link{mxLSTM}}}

\item{zoneout}{see \code{\link{mxLSTM}}}

\item{batchNorm}{see \code{\link{mxLSTM}}}

\item{activation}{activation function for update layers. "relu" or "tanh"}
}
\value{
mxSymbol
}
\description{
constructs an lstm cell. Identical to mxnet:::lstm, except for relu activation instead of tanh
}
