% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx.callback.log.R
\name{mx.callback.log}
\alias{mx.callback.log}
\title{mx.callback.log}
\usage{
mx.callback.log(period, loggerEnv)
}
\arguments{
\item{period}{interval of logger execution in epochs.}

\item{loggerEnv}{environment for storing the logger results}
}
\value{
The logger will create a list named 'logger' in loggerEnv 
        with the following items:
        \itemize{
          \item{train} {numeric vector with training metric evaluations}
          \item{eval}  {numeric vector with test metric evaluations}
          \item{checkpoint} {List of model weights for restoring the 
                             state after the specified epoch. 
                             See ?getMxNetModel on how to restore checkpoint}
         }
}
\description{
custom logger function that is executed after each 
             epoch of mxnet training.
}
\seealso{
\code{\link{plot_trainHistory}}
}
