% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxLSTMsource.R
\name{mxLSTMtrain}
\alias{mxLSTMtrain}
\title{mxLSTMtrain
#' @param model mxSymbol as returned by mxLSTMcreate}
\usage{
mxLSTMtrain(datTrain, datEval, symbol, batchSize, num.hidden, num.rounds,
  optimizer = "rmsprop", learning.rate = data.table(epoch = 1L, lr = 0.002),
  initializer = mx.init.Xavier(), initialModel = NULL, shuffle = TRUE,
  gammaInit, epoch.end.callback = NULL, ...)
}
\arguments{
\item{datTrain}{list with entries 'data' and 'label'. 
'data' is a 3D array of dimension num.features:seq.length:number of observations
'label is a 3D array of dimension num.outputs:seq.length:number of observations}

\item{datEval}{similar to datTrain, but for evaluation instead of training.}

\item{symbol}{mxSymbol as returned by \code{\link{mxLSTMcreate}}}

\item{batchSize}{see \code{\link{mxLSTM}}}

\item{num.hidden}{see \code{\link{mxLSTM}}}

\item{num.rounds}{see num.epoch argument in \code{\link{mxLSTM}}}

\item{optimizer}{see \code{\link{mxLSTM}}}

\item{learning.rate}{data.table with two columns: 
\itemize{
   \item{\code{epoch}} (integer) tells from which epoch onwards this learningrate is active. 
   Has to start with 1
   \item{\code{lr}} (numeric) the actual value for the learningrate
}\code{epoch}}

\item{initializer}{see \code{\link{mxLSTM}}}

\item{initialModel}{see \code{\link{mxLSTM}}}

\item{shuffle}{see \code{\link{mxLSTM}}}

\item{gammaInit}{see \code{\link{mxLSTM}}}

\item{epoch.end.callback}{function to be called at the end of each epoch.}

\item{...}{further arguments passed to optimizer}
}
\value{
object of class 'mxLSTM' 
       list: 'model' is the actual symbol, 'arg.params' and 'aux.params' are the parameters,
        'log' is the training log, 'optimizerEnv' is an optional environment with optimizer parameters.
}
\description{
train the LSTM
}
