% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_goodnessOfFit.R
\name{plot_goodnessOfFit}
\alias{plot_goodnessOfFit}
\title{plot_GoodnessOfFit}
\usage{
plot_goodnessOfFit(predicted, observed, plotIt = TRUE, combine = TRUE)
}
\arguments{
\item{predicted}{vector of predicted values}

\item{observed}{vector of observed values}

\item{plotIt}{If TRUE, plots are printed, if FALSE, they are only returned.}

\item{combine}{If TRUE, all plots are combined before, if FALSE, a list of individual plots is returned.}
}
\value{
a plot (combine = TRUE) or a list of plots (combine = FALSE)
}
\description{
Function to plot some godness of fit indicators for regression results
}
\details{
For details on the goodness of fit measures, see \code{\link{calc_goodnessOfFit}}
The plots contain the following information:
\describe{
\item{\code{Predicted vs observed values}}{scatterplot of predicted over observed values. \cr 
                                    Blue line indicates perfect fit, red line indicates best 
                                    fit to data. For the meaning of performance measures, 
                                    see \code{\link{calc_goodnessOfFit}}.}
\item{\code{Distribution of residuals}}{density plot of the distribution of residuals (red area) 
                                 with best fit gaussian (black dashed line). 
                                 Blue line at x=0, red line for mean of residuals. 
                                 For the meaning of performance measures, 
                                 see \code{\link{calc_goodnessOfFit}}.}
\item{\code{qqplot}}{standard qqplot. See \code{\link[stats]{qqnorm}}.}
\item{\code{Residuals over observed values}}{Scatterplot of residuals over observed values. 
                                      Blue line at 0. Red line for best fit to data.}
\item{\code{Residuals over order}}{Scatterplot of residuals over order (in case of timeseries: time). 
                            Red area indicates the 3 sigma interval around the mean of the 
                            residuals.}
}
}
\examples{
test <- plot_goodnessOfFit(predicted = 1:10 + rnorm(10,0,0.3), observed = 1:10)
}
\seealso{
\code{\link{calc_goodnessOfFit}}
}
