% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trainHistory.R
\name{plot_trainHistory}
\alias{plot_trainHistory}
\title{plot_trainHistory}
\usage{
plot_trainHistory(models)
}
\arguments{
\item{models}{a named vector with models.Can either be a model object or a path to a saved model}
}
\value{
plot as returned by highcharter
}
\description{
Plots the performance after each training epoch for a model 
as returned by \code{\link{fitLSTMmodel}}
}
\seealso{
\code{\link{getTrainHistory}}
}
