% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAll.R
\name{predictAll}
\alias{predictAll}
\title{predictAll}
\usage{
predictAll(object, newdata, ...)
}
\arguments{
\item{object}{trained model as returned by \code{\link[caret]{train}}.}

\item{newdata}{data.frame with new x variables}

\item{...}{additional arguments to the \code{predict} function of the model.}
}
\value{
data.frame with the predictions
}
\description{
Function to predict all y variables from a neuralnet regression model.
}
\seealso{
\code{\link[caret]{predict.train}}
}
