% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictLSTMmodel.R
\name{predictLSTMmodel}
\alias{predictLSTMmodel}
\title{predictLSTMmodel}
\usage{
predictLSTMmodel(model, dat, fullSequence = TRUE)
}
\arguments{
\item{model}{LSTM model as returned by \code{\link{mxLSTM}} or \code{\link{fitLSTMmodel}}}

\item{dat}{input data as provided by \code{\link{transformLSTMinput}} in the 'x' element of the list.}

\item{fullSequence}{Boolean. If FALSE, only the last predicted element of a sequence is returned.
If TRUE, a prediction for each step in the sequence is returned.
If the model was trained with optimizeFullSequence = FALSE, 
this will be set to FALSE with a warning.}
}
\value{
data.frame with predictions.
}
\description{
get predictions for an LSTM recurrent network
}
\details{
the sequence length is inferred from the \code{model} argument.
}
\seealso{
\code{\link{mxLSTM}}, \code{\link{fitLSTMmodel}}, \code{\link{getLSTMmodel}}
}
