% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictPreProcessor.R
\name{predictPreProcessor}
\alias{predictPreProcessor}
\title{predictPreProcessor}
\usage{
predictPreProcessor(preProcess, dat)
}
\arguments{
\item{preProcess}{Stored preProcessing procedure. Result of \code{\link{getPreProcessor}}.}

\item{dat}{data.frame containing the data to be preprocessed. Must have the correct column names.}
}
\value{
data.frame with preprocessed data
}
\description{
Perform stored preProcessing procedure on new data
}
\seealso{
\code{\link{getPreProcessor}}, \code{\link{invertPreProcessor}}
}
