% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreLSTMcheckpoint.R
\name{restoreLstmCheckpoint}
\alias{restoreLstmCheckpoint}
\title{restoreLstmCheckpoint}
\usage{
restoreLstmCheckpoint(model, checkpointNumber = NULL)
}
\arguments{
\item{model}{the model as returned by \code{\link{fitLSTMmodel}} or \code{\link{mxLSTM}}. 
Is returned as model$finalModel by \code{\link[caret]{train}}}

\item{checkpointNumber}{The number of the checkpoint to restore. 
Investigate the training history with 
\code{\link{plot_trainHistory}} to choose a checkpoint.}
}
\value{
the updated model. Parameters are set to the checkpoint and 
        model$log$activeCheckpoint is updated.
}
\description{
function to restore logged checkpoint from mxLSTM model
}
\seealso{
\code{\link{plot_trainHistory}}
}
