% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLstmModel.R
\name{saveLstmModel}
\alias{saveLstmModel}
\title{saveLstmModel}
\usage{
saveLstmModel(model, outFolder, overwrite = FALSE, fullHistory = TRUE)
}
\arguments{
\item{model}{the model as returned from \code{\link{mxLSTM}}}

\item{outFolder}{name of the folder to store results in.}

\item{overwrite}{If FALSE, function throws an error if outFolder already exists.}

\item{fullHistory}{Boolean. If TRUE, saves all the history, including logs of parameters.
If FALSE, only save bare last model (for storage efficiency)}
}
\value{
No return
}
\description{
saves an LSTM model that was trained using \code{\link{mxLSTM}}.
}
\seealso{
\code{\link{loadLstmModel}}
}
