% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformLSTMinput.R
\name{transformLSTMinput}
\alias{transformLSTMinput}
\title{transformLSTMinput}
\usage{
transformLSTMinput(dat, targetColumn, seq.length, seq.freq = seq.length)
}
\arguments{
\item{dat}{data.frame with variables. Is assumed to be a time-series without gaps. 
Each column corresponds to one variable}

\item{targetColumn}{Name of the column that represents the regression target. 
All other columns are considered as inputs}

\item{seq.length}{sequence length.}

\item{seq.freq}{frequency of sequence starts. Defaults to sequence length.}
}
\value{
a list of two: 'x' is an array with input variables as required for \code{\link{mxLSTM}}.
                       'y' is an array of target values as required for \code{\link{mxLSTM}}.
}
\description{
transforms a data.frame input into the correct format for the LSTM model
}
\details{
The last sequence is discarded if it is not complete. 
         Sequences containing missing values are discarded.
}
