% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxLSTMsource.R
\name{zoneout}
\alias{zoneout}
\title{zoneout}
\usage{
zoneout(thisState, prevState, p)
}
\arguments{
\item{thisState}{the current state (mx.symbol)}

\item{prevState}{the state from the previous sequence (mx.symbol).}

\item{p}{zoneout probability in (0;1]. Rounded to 6 digits}
}
\value{
During training: out(t) = x(t) * thisState(t) + (1 - x(t)) * prevState(t). 
        x(t) is a 0/1 mask for each element t with the probability p of being 0.
        During application: out(t) = (1 - p) * thisState(t) + p * prevState(t)
}
\description{
Applies zoneout as described in Krueger et al. 2017
}
