% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxLSTMsource.R
\name{mxLSTMcreate}
\alias{mxLSTMcreate}
\title{mxLSTMcreate}
\usage{
mxLSTMcreate(seq.length, num.outputs, num.hidden, optimizeFullSequence,
  dropoutLstm = 0, zoneoutLstm = 0, batchNormLstm = FALSE,
  batch.size = 128, activation = "relu")
}
\arguments{
\item{seq.length}{see \code{\link{mxLSTM}}}

\item{num.outputs}{number of final target variables}

\item{num.hidden}{see \code{\link{mxLSTM}}}

\item{optimizeFullSequence}{see \code{\link{mxLSTM}}}

\item{dropoutLstm}{see \code{\link{mxLSTM}}}

\item{zoneoutLstm}{see \code{\link{mxLSTM}}}

\item{batchNormLstm}{see \code{\link{mxLSTM}}}

\item{batch.size}{see \code{\link{mxLSTM}}}

\item{activation}{see \code{\link{mxLSTM}}}
}
\value{
MXSymbol
}
\description{
Creates the basic network.
 consists only of symbols, no binding to values yet.
}
