% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxLSTMsource.R
\name{mxLSTMsetup}
\alias{mxLSTMsetup}
\title{mxLSTMsetup}
\usage{
mxLSTMsetup(model, num.features, num.outputs, num.hidden, optimizeFullSequence,
  seq.length, batch.size, initializer, initialModel = NULL, gammaInit)
}
\arguments{
\item{model}{mxSymbol as returned by mxLSTMcreate}

\item{num.features}{number of input features.}

\item{num.outputs}{number of target variables}

\item{num.hidden}{see \code{\link{mxLSTM}}}

\item{optimizeFullSequence}{see \code{\link{mxLSTM}}}

\item{seq.length}{see \code{\link{mxLSTM}}}

\item{batch.size}{see \code{\link{mxLSTM}}}

\item{initializer}{see \code{\link{mxLSTM}}}

\item{initialModel}{see \code{\link{mxLSTM}}}

\item{gammaInit}{see \code{\link{mxLSTM}}}
}
\value{
MXExecutor
}
\description{
initialize the weights and input matrices with random number to create an executable model
}
